backend/requirements.txt
Flask
Flask-CORS
yfinance
pandas
numpy
requests
python-dotenv
gunicorn
crewai
groq
langchain
langchain-groq
duckduckgo-search
beautifulsoup4
lxml